TryOut Visualization of Agents and Pheromones
This is just a first try out, the code will have errors sometimes by running it and I'm sure we will not use that code in the end.
Still, for me it was really helpfull to understand what classes and methods we need and where things propably will become difficult.
I tried to add all my thoughts as comments, please read them if you running the code. Also I added a picture what the second plot in the code can look like
the lighter the dots are the more pheromones are at that positon.
I'm happy to read your ideas and thoughts!